Added lineHeight support
This is a bit of a complicated issue as line height only really applys to mulitple lines and with regards to canvas we need the first line to be set as the initial height.  For example when line-height is zero this should mean that all lines overlap each other, but if we set all line heights to zero, then we will not show any text.  So basically we must treat the first line as its rendered height, and then each subsequent line is the same plus the line height.  This doesnt really work well with varying height lines, but in the context of a line wrapped block of text all at the same font size, it works well.  This most likely also doesnt play nicely with valign…